+++ Sources/Tabman/Bar/BarButton/TMBarButton.swift
@@ -98,6 +98,11 @@ open class TMBarButton: UIControl {
         
         layout(in: contentView)
         layoutBadge(badge, in: contentView)
+
+        accessibilityLabel = item.accessibilityLabel ?? item.title
+        accessibilityHint = item.accessibilityHint
+        accessibilityTraits = [.button]
+        isAccessibilityElement = true
     }
     
     // MARK: Layout
@@ -158,6 +163,15 @@ open class TMBarButton: UIControl {
         let alpha = minimumAlpha + (selectionState.rawValue * minimumAlpha)
         
         self.alpha = alpha
+
+        switch selectionState {
+        case .selected:
+            accessibilityTraits.insert(.selected)
+        case .unselected:
+            accessibilityTraits.remove(.selected)
+        default:
+            break
+        }
     }
 }